Cleanup implementation details to prepare for a PR
